requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
coverage==6.3.2
Django==4.0.3
djangorestframework==3.13.1
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
yarl==1.7.2
